requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
